refactor: folder structure for NotFound page
