+++ src/pages/not-found/NotFound.scss
@@ -1,4 +1,4 @@
-@import '../styles/reset', '../styles/variables';
+@import '../../styles/reset', '../../styles/variables';
 
 .not-found--layout {
   background-color: $accent-background-thick;
